[cleanup] add actions and reducer
