+++ packages/waffle-chart/src/CdcWaffleChart.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from 'react'
+import React, { useCallback, useEffect, useState, useReducer } from 'react'
 import parse from 'html-react-parser'
 import { Group } from '@visx/group'
 import { Circle, Bar } from '@visx/shape'
@@ -21,7 +21,7 @@ import useDataVizClasses from '@cdc/core/helpers/useDataVizClasses'
 import coveUpdateWorker from '@cdc/core/helpers/coveUpdateWorker'
 
 import { Config } from './types/Config'
-import { useWaffleChart } from './hooks/useWaffleChart'
+import chartReducer from './store/chart.reducer'
 
 import './scss/main.scss'
 
@@ -31,12 +31,29 @@ type CdcWaffleChartProps = {
   isDashboard?: boolean
   isEditor?: boolean
   link?: string
-  setConfig?: Function
+  setConfig?: () => void
 }
 
 const WaffleChart = ({ config, isEditor, link = '' }) => {
   const { title, theme, shape, nodeWidth, nodeSpacer, prefix, suffix, subtext, content, orientation, filters, dataColumn, dataFunction, dataConditionalColumn, dataConditionalOperator, dataConditionalComparate, customDenom, dataDenom, dataDenomColumn, dataDenomFunction, roundToPlace } = config
-  const { themeColor, gaugeHeight, gaugeWidth, gaugeColor, dataFontSize, handleWaffleChartAriaLabel } = useWaffleChart(config)
+
+  const handleWaffleChartAriaLabel = (state, testing = false): string => {
+    // eslint-disable-next-line no-console
+    if (testing) console.log(`handleWaffleChartAriaLabels Testing On:`, state)
+    try {
+      let ariaLabel = 'Waffle chart'
+      if (state.title) {
+        ariaLabel += ` with the title: ${state.title}`
+      }
+      return ariaLabel
+    } catch (e) {
+      // eslint-disable-next-line no-console
+      console.error(e.message)
+    }
+  }
+
+  const gaugeColor = config.visual.colors[config.theme]
+  let dataFontSize = config.fontSize ? { fontSize: config.fontSize + 'px' } : null
 
   const calculateData = useCallback(() => {
     //If either the column or function aren't set, do not calculate
@@ -223,7 +240,7 @@ const WaffleChart = ({ config, isEditor, link = '' }) => {
         shape: shape,
         x: calculatePos(shape, 'x', i, nodeWidthNum, nodeSpacerNum),
         y: calculatePos(shape, 'y', i, nodeWidthNum, nodeSpacerNum),
-        color: themeColor[theme],
+        color: config.visual.colors[theme],
         opacity: i + 1 > 100 - Math.round(dataPercentage) ? 1 : 0.35
       }
       waffleData.push(newNode)
@@ -256,7 +273,7 @@ const WaffleChart = ({ config, isEditor, link = '' }) => {
 
   const xScale = scaleLinear({
     domain: [0, waffleDenominator],
-    range: [0, gaugeWidth]
+    range: [0, config.gauge.width]
   })
 
   return (
@@ -278,10 +295,10 @@ const WaffleChart = ({ config, isEditor, link = '' }) => {
                     {suffix ? suffix + ' ' : ' '} {config.valueDescription} {config.showDenominator && waffleDenominator ? waffleDenominator : ' '}
                   </div>
                   <div className='cove-waffle-chart__data--text'>{parse(content)}</div>
-                  <svg height={gaugeHeight} width={'100%'}>
+                  <svg height={config.gauge.height} width={'100%'}>
                     <Group>
-                      <foreignObject style={{ border: '1px solid black' }} x={0} y={0} width={gaugeWidth} height={gaugeHeight} fill='#fff' />
-                      <Bar x={0} y={0} width={xScale(waffleNumerator)} height={gaugeHeight} fill={gaugeColor} />
+                      <foreignObject style={{ border: '1px solid black' }} x={0} y={0} width={config.gauge.width} height={config.gauge.height} fill='#fff' />
+                      <Bar x={0} y={0} width={xScale(waffleNumerator)} height={config.gauge.height} fill={gaugeColor} />
                     </Group>
                   </svg>
                   <div className={'cove-waffle-chart__subtext subtext'}>{parse(subtext)}</div>
@@ -321,12 +338,8 @@ const WaffleChart = ({ config, isEditor, link = '' }) => {
 
 const CdcWaffleChart = ({ configUrl, config: configObj, isDashboard = false, isEditor = false, setConfig: setParentConfig }: CdcWaffleChartProps) => {
   // Default States
-  const [config, setConfig] = useState({ ...defaults })
-  const [loading, setLoading] = useState(true)
-
-  const [currentViewport, setCurrentViewport] = useState('lg')
-  const [coveLoadedHasRan, setCoveLoadedHasRan] = useState(false)
-  const [container, setContainer] = useState()
+  const [state, dispatch] = useReducer(chartReducer, { config: configObj ?? defaults, loading: true, preview: false, viewport: 'lg', coveLoadedHasRan: false, container: null })
+  const { loading, config, viewport: currentViewport, coveLoadedHasRan, container } = state
 
   // Default Functions
   const updateConfig = newConfig => {
@@ -335,12 +348,10 @@ const CdcWaffleChart = ({ configUrl, config: configObj, isDashboard = false, isE
         newConfig[key] = { ...defaults[key], ...newConfig[key] }
       }
     })
-
     newConfig.runtime = {}
     newConfig.runtime.uniqueId = Date.now()
-
     newConfig.runtime.editorErrorMessage = ''
-    setConfig(newConfig)
+    dispatch({ type: 'SET_CONFIG', payload: newConfig })
   }
 
   const loadConfig = useCallback(async () => {
@@ -355,7 +366,7 @@ const CdcWaffleChart = ({ configUrl, config: configObj, isDashboard = false, isE
 
     const processedConfig = { ...(await coveUpdateWorker(response)) }
     updateConfig({ ...defaults, ...processedConfig })
-    setLoading(false)
+    dispatch({ type: 'SET_LOADING', payload: false })
   }, [])
 
   // Custom Functions
@@ -364,8 +375,7 @@ const CdcWaffleChart = ({ configUrl, config: configObj, isDashboard = false, isE
   const resizeObserver = new ResizeObserver(entries => {
     for (let entry of entries) {
       let newViewport = getViewport(entry.contentRect.width * 2) // Data bite is usually presented as small, so we scale it up for responsive calculations
-
-      setCurrentViewport(newViewport)
+      dispatch({ type: 'SET_VIEWPORT', payload: newViewport })
     }
   })
 
@@ -373,23 +383,25 @@ const CdcWaffleChart = ({ configUrl, config: configObj, isDashboard = false, isE
     if (node !== null) {
       resizeObserver.observe(node)
     }
-    setContainer(node)
+    dispatch({ type: 'SET_CONTAINER', payload: node })
   }, [])
 
   //Load initial config
   useEffect(() => {
+    // eslint-disable-next-line no-console
     loadConfig().catch(err => console.log(err))
   }, [])
 
   useEffect(() => {
     if (config && !coveLoadedHasRan && container) {
       publish('cove_loaded', { config: config })
-      setCoveLoadedHasRan(true)
+      dispatch({ type: 'SET_COVE_LOADED_HAS_RAN', payload: true })
     }
   }, [config, container])
 
   //Reload config if config object provided/updated
   useEffect(() => {
+    // eslint-disable-next-line no-console
     loadConfig().catch(err => console.log(err))
   }, [])
